null out temp_buffer on deinit
denitNoContext initNoContext checks if temp_buffer is null to know whether it needs to initialize it, so temp_buffer should be nulled out in deinit to support reinitializing the lib.

The use case for this is "demo list" style modes of app frameworks where you may initialize and run a sequence of demo applications for testing or as a feature demonstration, and multiple of these demo apps may want to initialize and use zgui, and then deinitialize it.