make explicitly named param structs pub
If there are enough params that a separately named struct decl is needed, it's likely a widget type that you'll want to make wrapper widgets for, which means it's helpful to be able to declare a local of the param struct.

I originally was doing this piecemeal as I needed, this change goes through all the currently exposed widget functions. Params that use an inline struct decl in the params didn't get the same treatment, but maybe worth considering for some?